feat: nuevos exports
