+++ lib/ChoquesPermitidos.js
@@ -1,5 +1,19 @@
+/*
+Copyright (c) 2021, Andrés Urmeneta B. <aurmeneta@uc.cl>
+Permission to use, copy, modify, and/or distribute this software for any
+purpose with or without fee is hereby granted, provided that the above
+copyright notice and this permission notice appear in all copies.
+THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+*/
+
 class ChoquesPermitidos {
-  constructor () {
+  constructor() {
     this.choques = []
   }
 
@@ -12,7 +26,7 @@ class ChoquesPermitidos {
      * @param tipo2 Tipo de módulo
      * @param permitido Boolean si el choque está permitido o no.
      */
-  anadirChoque (sigla1, tipo1, sigla2, tipo2, permitido) {
+  anadirChoque(sigla1, tipo1, sigla2, tipo2, permitido) {
     // Comprobar si es un choque especifico para dos siglas o general para varias siglas.
     const choque = this.buscarChoque(sigla1, tipo1, sigla2, tipo2)
 
@@ -29,7 +43,7 @@ class ChoquesPermitidos {
      * @param tipo2 Tipo de módulo
      * @returns {Choque}
      */
-  buscarChoque (sigla1, tipo1, sigla2, tipo2) {
+  buscarChoque(sigla1, tipo1, sigla2, tipo2) {
     const choques = this.buscarChoques(sigla1, tipo1, sigla2, tipo2)
 
     if (choques.length > 1) throw new ExcepcionMultiplesResultados(choques)
@@ -44,14 +58,14 @@ class ChoquesPermitidos {
      * @param tipo2 Tipo de módulo.
      * @returns {*[]}
      */
-  buscarChoques (sigla1, tipo1, sigla2, tipo2) {
+  buscarChoques(sigla1, tipo1, sigla2, tipo2) {
     return this.choques
       .filter(choque => {
         return (choque.sigla1 === sigla1 || choque.sigla1 === '*') && (choque.tipo1 === tipo1 || choque.tipo1 === '*') &&
-                    (choque.sigla2 === sigla2 || choque.sigla2 === '*') && (choque.tipo2 === tipo1 || choque.tipo2 === '*') ||
-                    // Comprobar también si los argumentos se dieron en el orden inverso.
-                    (choque.sigla1 === sigla2 || choque.sigla1 === '*') && (choque.tipo1 === tipo2 || choque.tipo1 === '*') &&
-                    (choque.sigla2 === sigla1 || choque.sigla2 === '*') && (choque.tipo1 === tipo2 || choque.tipo1 === '*')
+          (choque.sigla2 === sigla2 || choque.sigla2 === '*') && (choque.tipo2 === tipo1 || choque.tipo2 === '*') ||
+          // Comprobar también si los argumentos se dieron en el orden inverso.
+          (choque.sigla1 === sigla2 || choque.sigla1 === '*') && (choque.tipo1 === tipo2 || choque.tipo1 === '*') &&
+          (choque.sigla2 === sigla1 || choque.sigla2 === '*') && (choque.tipo1 === tipo2 || choque.tipo1 === '*')
       })
   }
 
@@ -63,7 +77,7 @@ class ChoquesPermitidos {
      * @param tipo2 Tipo de módulo
      * @returns {boolean}
      */
-  evaluarChoque (sigla1, tipo1, sigla2, tipo2) {
+  evaluarChoque(sigla1, tipo1, sigla2, tipo2) {
     // Busca el choque en los arreglos de choques permitidos.
     const choque = this.buscarChoque(sigla1, tipo1, sigla2, tipo2)
 
@@ -72,16 +86,16 @@ class ChoquesPermitidos {
   }
 }
 
-function ExcepcionMultiplesResultados (resultados) {
+function ExcepcionMultiplesResultados(resultados) {
   this.resultados = resultados
 
   this.toString = () => {
     return 'Múltiples choques encontrados. Revisa los choques permitidos o los realiza una búsqueda más precisa.' +
-            `Se han encontrado ${resultados.length} resultados.`
+      `Se han encontrado ${resultados.length} resultados.`
   }
 }
 
-function Choque (sigla1, tipo1, sigla2, tipo2, permitido) {
+function Choque(sigla1, tipo1, sigla2, tipo2, permitido) {
   this.sigla1 = sigla1
   this.tipo1 = tipo1
   this.sigla2 = sigla2
