+++ lib/Curso.js
@@ -1,8 +1,22 @@
+/*
+Copyright (c) 2021, Andrés Urmeneta B. <aurmeneta@uc.cl>
+Permission to use, copy, modify, and/or distribute this software for any
+purpose with or without fee is hereby granted, provided that the above
+copyright notice and this permission notice appear in all copies.
+THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+*/
+
 const Modulo = require('./Modulo')
 
 class Curso {
   // TODO: Forzar tipos de los parámetros.
-  constructor (nrc, sigla, permiteRetiro, ingles, seccion, aprobacionEspecial, area, formato, categoria, nombre,
+  constructor(nrc, sigla, permiteRetiro, ingles, seccion, aprobacionEspecial, area, formato, categoria, nombre,
     profesor, campus, creditos, vacantesDisponibles, horario) {
     this.nrc = nrc
     this.sigla = sigla
@@ -27,7 +41,7 @@ class Curso {
      * @param curso2
      * @returns {boolean}
      */
-  static mismoHorario (curso1, curso2) {
+  static mismoHorario(curso1, curso2) {
     const horario1 = curso1.horario
     const horario2 = curso2.horario
 
@@ -47,7 +61,7 @@ class Curso {
      * @param choquesPermitidos
      * @returns {boolean}
      */
-  static horariosCompatibles (curso1, curso2, choquesPermitidos) {
+  static horariosCompatibles(curso1, curso2, choquesPermitidos) {
     const horario1 = curso1.horario
     const horario2 = curso2.horario
     const sigla1 = curso1.sigla
@@ -73,7 +87,7 @@ class Curso {
      * @param horario
      * @returns {Curso}
      */
-  static cursoMinimo (nrc, sigla, seccion, nombre, profesor, vacantesDisponibles, horario) {
+  static cursoMinimo(nrc, sigla, seccion, nombre, profesor, vacantesDisponibles, horario) {
     return new Curso(nrc, sigla, undefined, undefined, seccion, undefined,
       '', '', '', nombre, profesor, '', 0, vacantesDisponibles, horario)
   }
