+++ lib/Modulo.js
@@ -1,5 +1,20 @@
+/*
+Copyright (c) 2021, Andrés Urmeneta B. <aurmeneta@uc.cl>
+Permission to use, copy, modify, and/or distribute this software for any
+purpose with or without fee is hereby granted, provided that the above
+copyright notice and this permission notice appear in all copies.
+THE SOFTWARE IS PROVIDED "AS IS" AND THE AUTHOR DISCLAIMS ALL WARRANTIES
+WITH REGARD TO THIS SOFTWARE INCLUDING ALL IMPLIED WARRANTIES OF
+MERCHANTABILITY AND FITNESS. IN NO EVENT SHALL THE AUTHOR BE LIABLE FOR
+ANY SPECIAL, DIRECT, INDIRECT, OR CONSEQUENTIAL DAMAGES OR ANY DAMAGES
+WHATSOEVER RESULTING FROM LOSS OF USE, DATA OR PROFITS, WHETHER IN AN
+ACTION OF CONTRACT, NEGLIGENCE OR OTHER TORTIOUS ACTION, ARISING OUT OF
+OR IN CONNECTION WITH THE USE OR PERFORMANCE OF THIS SOFTWARE.
+*/
+
 class Modulo {
-  static get TIPOS () {
+  // TODO: getter sin sentido acá
+  static get TIPOS() {
     return [
       'CLAS',
       'LAB',
@@ -12,7 +27,7 @@ class Modulo {
       'TES']
   }
 
-  static get DIAS () {
+  static get DIAS() {
     return [
       'L',
       'M',
@@ -22,12 +37,12 @@ class Modulo {
       'S']
   }
 
-  static get MODULOS () {
+  static get MODULOS() {
     return 8
   }
 
   // TODO: Forzar tipo de los parámetros.
-  constructor (tipo, dia, modulo, sala) {
+  constructor(tipo, dia, modulo, sala) {
     this.tipo = tipo
     this.dia = dia
     this.modulo = modulo
@@ -35,11 +50,11 @@ class Modulo {
     this.hora = modulo
   }
 
-  static modulosIdenticos (a, b) {
+  static modulosIdenticos(a, b) {
     return a.tipo === b.tipo && a.dia === b.dia && a.modulo === b.modulo
   }
 
-  static modulosCompatibles (a, b) {
+  static modulosCompatibles(a, b) {
     return (a.dia !== b.dia) || (a.modulo !== b.modulo) || (a.dia === 'SIN HORARIO') || (b.dia === 'SIN HORARIO')
   }
 }
